projet_MyDiagAI/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
pymysql==1.1.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
scikit-learn==1.3.0